+++ hooks/src/lib/useCardArrayValidate/useCardArrayValidate.ts
