+++ hooks/src/lib/useCardHolderName/useCardHolderName.ts
@@ -1,29 +1,23 @@
-import useCardValidate from './useCardValidate';
+import useCardValidate from '../useCardValidate/useCardValidate';
 
-const useCardPassword = (initValue: string, maxLength: number = 2) => {
+const useCardHolderName = (initValue: string, maxLength: number = 50) => {
   const validateOnChange = (newValue: string) => {
     if (newValue.length > maxLength) {
       return {
         isValid: false,
-        errorMessage: `비밀번호는 ${maxLength}글자 까지만 입력이 가능해요.`,
+        errorMessage: `이름은 ${maxLength}글자 까지만 입력이 가능해요.`,
       };
     }
-    if (!/^\d*$/.test(newValue)) {
+    if (!/^\[a-zA-Z]*$/.test(newValue)) {
       return {
         isValid: false,
-        errorMessage: '비밀번호는 숫자만 입력이 가능해요.',
+        errorMessage: '이름은 영어만 입력이 가능해요.',
       };
     }
     return { isValid: true, errorMessage: '' };
   };
 
   const validateOnBlur = () => {
-    if (value.length !== maxLength) {
-      return {
-        isValid: false,
-        errorMessage: `비밀번호는 ${maxLength}글자로 입력해 주세요.`,
-      };
-    }
     return { isValid: true, errorMessage: '' };
   };
 
@@ -49,4 +43,4 @@ const useCardPassword = (initValue: string, maxLength: number = 2) => {
     onFocusHandler,
   };
 };
-export default useCardPassword;
+export default useCardHolderName;
