+++ hooks/src/lib/useCardPassword/useCardPassword.ts
@@ -1,27 +1,27 @@
-import useCardArrayValidate from './useCardArrayValidate';
+import useCardValidate from '../useCardValidate/useCardValidate';
 
-const useCardNumbers = (initValue: string[], maxLength: number = 4) => {
+const useCardPassword = (initValue: string, maxLength: number = 2) => {
   const validateOnChange = (newValue: string) => {
     if (newValue.length > maxLength) {
       return {
         isValid: false,
-        errorMessage: `카드번호는 ${maxLength}글자 까지만 입력이 가능해요.`,
+        errorMessage: `비밀번호는 ${maxLength}글자 까지만 입력이 가능해요.`,
       };
     }
     if (!/^\d*$/.test(newValue)) {
       return {
         isValid: false,
-        errorMessage: '카드번호는 숫자만 입력이 가능해요.',
+        errorMessage: '비밀번호는 숫자만 입력이 가능해요.',
       };
     }
     return { isValid: true, errorMessage: '' };
   };
 
   const validateOnBlur = () => {
-    if (value.join('').length !== maxLength * initValue.length) {
+    if (value.length !== maxLength) {
       return {
         isValid: false,
-        errorMessage: `카드번호는 ${maxLength * initValue.length}글자로 입력해 주세요.`,
+        errorMessage: `비밀번호는 ${maxLength}글자로 입력해 주세요.`,
       };
     }
     return { isValid: true, errorMessage: '' };
@@ -34,7 +34,7 @@ const useCardNumbers = (initValue: string[], maxLength: number = 4) => {
     onChangeHandler,
     onFocusHandler,
     onBlurHandler,
-  } = useCardArrayValidate({
+  } = useCardValidate({
     initValue,
     validateOnChange,
     validateOnBlur,
@@ -49,4 +49,4 @@ const useCardNumbers = (initValue: string[], maxLength: number = 4) => {
     onFocusHandler,
   };
 };
-export default useCardNumbers;
+export default useCardPassword;
